Implement `Sweep` for `Face`
