+++ crates/fj-kernel/src/algorithms/sweep.rs
@@ -1,5 +1,6 @@
 //! Sweeping objects along a path to create new objects
 
+mod face;
 mod sketch;
 
 use fj_interop::mesh::Color;
